refactor: 💡 renamed edit-profile to edit-validator-profile
