+++ components/overview/index.js
@@ -53,6 +53,7 @@ const Overview = () => {
 		if (get(stashAccount, 'address')) {
 			const kusamaAddress = encodeAddress(decodeAddress(stashAccount.address), 2);
 			axios.get(`user/${kusamaAddress}`).then(({ data }) => {
+				console.log(data);
 				if (data.message === 'No data found!') setError(true);
 				setUserData(data);
 			}).catch(() => {
